fix assert

+++ src/Hipace.cpp
@@ -144,7 +144,7 @@ Hipace::Hipace () :
     MakeGeometry();
 
     AMREX_ALWAYS_ASSERT_WITH_MESSAGE(
-        (((m_comms_buffer_max_leading_slices + m_comms_buffer_max_trailing_slices)
+        (((double(m_comms_buffer_max_leading_slices) + m_comms_buffer_max_trailing_slices)
         * amrex::ParallelDescriptor::NProcs()) > m_3D_geom[0].Domain().length(2))
         || (m_max_step <= amrex::ParallelDescriptor::NProcs()),
         "comms_buffer_max_leading_slices and comms_buffer_max_trailing_slices must be large enough"
